+++ stack/source/nwk_interface/protocol_timer.h
@@ -31,7 +31,7 @@ typedef struct {
     uint16_t time_drifts;
 } protocol_timer_t;
 
-void protocol_timer_cb(uint16_t ticks);
+void protocol_timer_cb(int ticks);
 void protocol_timer_start(protocol_timer_id_e id, void (*passed_fptr)(uint16_t), uint32_t time_ms);
 void protocol_timer_stop(protocol_timer_id_e id);
